requirements.txt
accelerate==1.7.0
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.3.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
deprecation==2.1.0
fastapi==0.115.13
filelock==3.18.0
fsspec==2025.5.1
getdaft==0.4.18
grpcio==1.73.0
grpcio-health-checking==1.73.0
grpcio-tools==1.73.0
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
validators==0.34.0
weaviate-client==4.15.2
